created bottom level as

+++ Common/UploadBuffer.h
@@ -22,7 +22,7 @@ public:
             mElementByteSize = d3dUtil::CalcConstantBufferByteSize(sizeof(T));
 
         auto heapProperties = CD3DX12_HEAP_PROPERTIES(D3D12_HEAP_TYPE_UPLOAD);
-        auto resourceDesc = CD3DX12_RESOURCE_DESC::Buffer(mElementByteSize * elementCount);
+        auto resourceDesc = CD3DX12_RESOURCE_DESC::Buffer(static_cast<UINT64>(mElementByteSize) * elementCount);
         ThrowIfFailed(device->CreateCommittedResource(
             &heapProperties,
             D3D12_HEAP_FLAG_NONE,
